+++ changelog.md
@@ -1,8 +1,18 @@
+# Version 0.11.0.0 (2023-08-29)
+
+- Add support for GHC 9.8.
+
+- Re-export functionality relating to GHC's constraint solving `TcS` monad,
+  such as `{get,set}InertSet`, `{get,set}TcEvBindsMap`.
+
+- Re-export `readTcRef` and `writeTcRef`.
+
 # Version 0.10.0.0 (2023-02-28)
 
 - Introduce `resolveImport`, and make `PkgQual` opaque.
+
 - Rename `tcRewriterWanteds` to `tcRewriterNewWanteds`
-  (bringing it in line with nomenclature in ghc 9.4).
+  (bringing it in line with nomenclature in GHC 9.4).
 
 # Version 0.9.0.0 (2023-01-24)